requirements.txt
locust
pyyaml
python-dotenv